[Feat] Create New Dashboard For Admins
Why: some features only admin can manage
How: as Dashboard , shared component
