+++ src/App.jsx
@@ -34,7 +34,7 @@ import UserProfile from "./components/Dashboard/Profile/UserProfile";
 import InternshipResponse from "./components/Dashboard/FormData/InternshipResponse";
 import EditPublicProfile from "./components/Dashboard/Profile/EditPublicProfile/EditPublicProfile";
 import Volunteer from "./components/Opportunities/Volunteer/Volunteer";
-import TimeLineEvent from "./components/Opportunities/TimeLineEvent/TimeLineEvent";
+import DisplayCommunityEvents from "./components/Opportunities/DisplayCommunityEvents";
 import TheCyberXcel from "./components/Opportunities/TheCyberXcel/TheCyberXcel";
 import OpenSecProjects from "./components/Opportunities/OpenSecProjects/OpenSecProjects";
 import DashboardRoute from "./components/Dashboard/DashboardRoute";
@@ -53,6 +53,7 @@ import ExploreRoutes from "./components/Explore/ExploreRoutes";
 import Leaderboard from "./components/Other/CyberGames/Leaderboard/Leaderboard";
 import SettingsRoute from "./components/Dashboard/Settings";
 import CheatSheetsRoutes from "./components/CheatSheets/CheatSheetsRoutes";
+import AdminDashboardRoute from "./components/AdminDashboard/AdminDashboardRoute";
 // import isAuthenticated from "./features/isAuthenticated";
 // import ChatBot from "./components/ChatBot/ChatBot";
 
@@ -78,11 +79,14 @@ const App = () => {
 
     const hideHomeHeader = () => {
         const pathDashboard = pathname.includes("/dashboard");
+        const pathAdminDashboard = pathname.includes("/admin-dashboard");
         const pathLogin = pathname.includes("/login");
         const pathRegister = pathname.includes("/register");
         const pathForgetPassword = pathname.includes("/forgetPassword");
         const pathResetPassword = pathname.includes("/resetPassword");
-        return pathDashboard || pathLogin || pathRegister || pathForgetPassword || pathResetPassword;
+        return (
+            pathDashboard || pathAdminDashboard || pathLogin || pathRegister || pathForgetPassword || pathResetPassword
+        );
     };
 
     const [isOpen, setIsOpen] = useState(false);
@@ -133,7 +137,7 @@ const App = () => {
                         <Route path={"/dashboard/settings/*"} element={<SettingsRoute />} />
 
                         <Route path={"/dashboard/*"} element={<DashboardRoute />} />
-
+                        <Route path={"/admin-dashboard/*"} element={<AdminDashboardRoute />} />
                         <Route exact path={"/contact"} element={<ContactForm />} />
 
                         <Route>
@@ -151,7 +155,7 @@ const App = () => {
 
                         <Route exact path={"/volunteer"} element={<Volunteer />} />
                         <Route exact path={"/opensec-projects"} element={<OpenSecProjects />} />
-                        <Route exact path={"/timeline-events"} element={<TimeLineEvent />} />
+                        <Route exact path={"/community-events"} element={<DisplayCommunityEvents />} />
                         <Route exact path={"/thecyberxcel"} element={<TheCyberXcel />} />
                         <Route exact path={"/thecyberspeak"} element={<TheCyberXcel />} />
