+++ src/components/Header/UserOptions/UserOptions.jsx
@@ -5,6 +5,7 @@ import { FaUserAstronaut, FaUserCircle } from "react-icons/fa";
 import { BiLogOut } from "react-icons/bi";
 import { IoMdLogIn } from "react-icons/io";
 import { MdDashboard } from "react-icons/md";
+import { RiAdminFill } from "react-icons/ri";
 
 import { useDispatch, useSelector } from "react-redux";
 import { RouterLink, UserOptionsContainer, UserOptionsDropdownContainer } from "./UserOptionsElements";
@@ -56,6 +57,7 @@ const DropdownMenu = () => {
             </RouterLink>
         );
     };
+
     return (
         <UserOptionsDropdownContainer className={"dropdown"}>
             <div className={"menu"}>
@@ -70,6 +72,11 @@ const DropdownMenu = () => {
                     </>
                 ) : (
                     <>
+                        {true && ( // user.type === "admin"
+                            <DropdownItem to={"/admin-dashboard"} leftIcon={<RiAdminFill />}>
+                                Admin Dashboard
+                            </DropdownItem>
+                        )}
                         <DropdownItem to={"/dashboard"} leftIcon={<MdDashboard />}>
                             Dashboard
                         </DropdownItem>
@@ -77,7 +84,6 @@ const DropdownMenu = () => {
                             Public Profile
                         </DropdownItem>
                         {/* {!inDevelopment ? ( */}
-
                         {/* ) : null} */}
                         {/* <DropdownItem to={"/settings"} leftIcon={<FcSettings />}> */}
                         {/*    Settings */}
